# openapi.yaml
openapi: 3.1.0
info:
  title: DOGFOOD Portal API
  version: 1.0.0
  description: >-
    Local-first hackathon submission, judging, community, and organizer API.
    Session identity is carried in an HttpOnly session cookie. Runtime data is in memory.
servers:
  - url: http://localhost:8080
security: []
paths:
...
  /api/organizer/event:
    put:
      summary: Configure event dates, tracks, prizes, and voting access
      security: [{ SessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventConfiguration" }
      responses:
        "200": { description: Updated event configuration }
        "403": { description: Organizer role required }
...
  /api/v1/judge-invitations/{token}/accept:
    post:
      summary: Accept an invited judge seat and issue a judge session
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Judge identity and session cookie issued }
        "404": { description: Invalid, expired, or used invitation }
  /api/v1/teams/{id}/invitations:
    post:
      summary: Create an email-bound, expiring team invite
      security: [{ SessionCookie: [] }]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
      responses:
        "201": { description: One-time invite token }
        "403": { description: Team ownership required }
...
  /api/v1/projects:
    get:
      summary: Search public projects
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: track, in: query, schema: { type: string } }
      responses:
        "200": { description: Filtered project list }
    post:
      summary: Submit a project for the participant's team
      security: [{ SessionCookie: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "201": { description: Project created }
        "403": { description: Participant role or open deadline required }
  /api/v1/projects/{id}:
    get:
      summary: Read a public project and community data
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200": { description: Project detail; judge results remain private before publication }
        "404": { description: Project not found }
    put:
      summary: Edit the participant's project before the deadline
      security: [{ SessionCookie: [] }]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "200": { description: Project updated }
        "403": { description: Team ownership or deadline check failed }
...
  /api/v1/projects/{id}/comments:
    post:
      summary: Add a rate-limited public comment
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                author: { type: string, maxLength: 80 }
                text: { type: string, maxLength: 2000 }
      responses:
        "201": { description: Comment recorded }
        "429": { description: Rate limit exceeded }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
    EventConfiguration:
      type: object
      properties:
        name: { type: string, maxLength: 120 }
        submissions_close: { type: [string, "null"], format: date-time }
        voting_open: { type: [string, "null"], format: date-time }
        voting_close: { type: [string, "null"], format: date-time }
        results_publish_at: { type: [string, "null"], format: date-time }
        voting_access: { type: string, enum: [open, email, authenticated] }
        tracks: { type: array, items: { type: object, required: [name], properties: { id: { type: string }, name: { type: string } } } }
        prizes: { type: array, items: { type: [string, object] } }
    ProjectInput:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 120 }
        track: { type: string }
        summary: { type: string, maxLength: 2000 }
        repo_url: { type: string, format: uri }